setup: cleanup and include gpiod.h in sdists
Remove author info, just leave maintainer name/email. It doesn't feel
quite right putting other names as the "author" without permission.
Bartosz is still mentionedin the copyright of all the gpiod code.

Add metadata for the URL and long description (README.md content) and
python_requires for >=3.

Also add a MANIFEST.in to include gpiod.h in sdist tarballs, and list it
as a dependency for the extension (which annoyingly isn't enough to get
setuptools to include it in sdist).